some partial stuff

+++ tests/toplevel_unit.c
@@ -86,13 +86,12 @@ void test_example() {
     const roaring_bitmap_t *allmybitmaps[] = {r1, r2, r3};
     roaring_bitmap_t *bigunion = roaring_bitmap_or_many(3, allmybitmaps);
     assert_true(roaring_bitmap_equals(r1_2_3, bigunion));
-    roaring_bitmap_t *bigunionheap = roaring_bitmap_or_many_heap(3, allmybitmaps);
+    roaring_bitmap_t *bigunionheap =
+        roaring_bitmap_or_many_heap(3, allmybitmaps);
     assert_true(roaring_bitmap_equals(r1_2_3, bigunionheap));
     roaring_bitmap_free(r1_2_3);
     roaring_bitmap_free(bigunion);
     roaring_bitmap_free(bigunionheap);
-
-
 
     // we can compute intersection two-by-two
     roaring_bitmap_t *i1_2 = roaring_bitmap_and(r1, r2);
@@ -1011,7 +1010,8 @@ void test_rand_flips() {
                 }
         }
         roaring_bitmap_run_optimize(r);
-        printf(" and actual card = %d\n", roaring_bitmap_get_cardinality(r));
+        printf(" and actual card = %d\n",
+               (int)roaring_bitmap_get_cardinality(r));
 
         for (int i = 0; i < flip_trials; ++i) {
             int start = rand() % (range - 1);
